MDL-67673 phpunit: Fix the return type of template methods
All the setup/teardown/pre/post/conditions template methods
now are required to return void. This was warned with phpunit 7
and now is enforced.

At the same time, fix a few wrong function names,
provider data and param types, return statements...

+++ grade/tests/grades_grader_gradingpanel_point_external_store_test.php
@@ -279,7 +279,7 @@ class store_test extends advanced_testcase {
      * @param int $maxvalue The max value of the forum
      * @param int $suppliedvalue The value that was submitted
      */
-    public function test_execute_store_out_of__range(int $maxvalue, int $suppliedvalue): void {
+    public function test_execute_store_out_of__range(int $maxvalue, float $suppliedvalue): void {
         $this->resetAfterTest();
 
         $forum = $this->get_forum_instance([
